+++ crypto/prepare_gdk_clang.sh
@@ -36,9 +36,9 @@ for arch in $all_archs; do
     mkdir -p ../src/main/jniLibs/$arch $PWD/gdk-android-jni$arch
     #./tools/build.sh --buildtype=debug --install $PWD/gdk-android-jni$arch --ndk $arch
     ./tools/build.sh --install $PWD/gdk-android-jni$arch --ndk $arch
-    cp gdk-android-jni$arch/lib/$arch/* ../src/main/jniLibs/$arch
-    cp gdk-android-jni$arch/java/com/blockstream/libgreenaddress/GDK.java ../src/main/java/com/blockstream/libgreenaddress/GDK.java
-    cp gdk-android-jni$arch/java/com/blockstream/libwally/Wally.java ../src/main/java/com/blockstream/libwally/Wally.java
+    cp gdk-android-jni$arch/lib/$arch/*.so ../src/main/jniLibs/$arch
+    cp gdk-android-jni$arch/share/java/com/blockstream/libgreenaddress/GDK.java ../src/main/java/com/blockstream/libgreenaddress/GDK.java
+    cp gdk-android-jni$arch/share/java/com/blockstream/libwally/Wally.java ../src/main/java/com/blockstream/libwally/Wally.java
 done
 
 cd ..