+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unrelease]
+
+### Fixed
+- Setup Pgp on multiple multisig networks
+
 ## [4.0.2] - 2023-04-20
 
 ### Fixed