+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Setup Pgp on multiple multisig networks
+- Sweep transaction in Watch-Only mode
 
 ## [4.0.2] - 2023-04-20